requirements.txt
Werkzeug>=2.0
Flask==2.3.3
python-dotenv>=1.0